requirements.txt
speechmatics-flow
pyaudio
python-dotenv
websockets==13.1
requests